requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
torch>=2.0
transformers>=4.50
pillow>=10.0
pypdfium2>=4.0
